Moved JsonConverters not specific to wallet to the Stratis.Bitcoin project
