+++ src/Stratis.Bitcoin.Features.Wallet.Tests/WalletManagerTest.cs
@@ -9,9 +9,9 @@ using NBitcoin;
 using Newtonsoft.Json;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Features.Wallet.Interfaces;
-using Stratis.Bitcoin.Features.Wallet.JsonConverters;
 using Stratis.Bitcoin.Tests.Logging;
 using Stratis.Bitcoin.Utilities;
+using Stratis.Bitcoin.Utilities.JsonConverters;
 using Xunit;
 
 namespace Stratis.Bitcoin.Features.Wallet.Tests
