+++ src/Stratis.Bitcoin.Features.WatchOnlyWallet/Models/WatchOnlyWalletModel.cs
@@ -4,7 +4,7 @@ using NBitcoin;
 using Newtonsoft.Json;
 using Stratis.Bitcoin.Features.RPC.Models;
 using Stratis.Bitcoin.Features.Wallet;
-using Stratis.Bitcoin.Features.Wallet.JsonConverters;
+using Stratis.Bitcoin.Utilities.JsonConverters;
 
 namespace Stratis.Bitcoin.Features.WatchOnlyWallet.Models
 {
