+++ src/Stratis.Bitcoin.Features.WatchOnlyWallet/WatchOnlyWallet.cs
@@ -5,7 +5,7 @@ using NBitcoin;
 using NBitcoin.JsonConverters;
 using Newtonsoft.Json;
 using Stratis.Bitcoin.Features.Wallet;
-using Stratis.Bitcoin.Features.Wallet.JsonConverters;
+using Stratis.Bitcoin.Utilities.JsonConverters;
 
 namespace Stratis.Bitcoin.Features.WatchOnlyWallet
 {
