+++ src/Stratis.Bitcoin.Tests/Utilities/NetworkHelpersTest.cs
@@ -1,54 +1,54 @@
 ﻿using System;
 using NBitcoin;
-using Stratis.Bitcoin.Features.Wallet.Helpers;
+using Stratis.Bitcoin.Utilities;
 using Xunit;
 
-namespace Stratis.Bitcoin.Features.Wallet.Tests
+namespace Stratis.Bitcoin.Tests.Utilities
 {
-    public class WalletHelpersTest
+    public class NetworkHelpersTest
     {
         [Fact]
         public void GetMainNetworkRetuirnsNetworkMain()
         {
-            Network network = WalletHelpers.GetNetwork("main");
+            Network network = NetworkHelpers.GetNetwork("main");
             Assert.Equal(Network.Main, network);
         }
 
         [Fact]
         public void GetMainNetNetworkRetuirnsNetworkMain()
         {
-            Network network = WalletHelpers.GetNetwork("mainnet");
+            Network network = NetworkHelpers.GetNetwork("mainnet");
             Assert.Equal(Network.Main, network);
         }
 
         [Fact]
         public void GetTestNetworkRetuirnsNetworkTest()
         {
-            Network network = WalletHelpers.GetNetwork("test");
+            Network network = NetworkHelpers.GetNetwork("test");
             Assert.Equal(Network.TestNet, network);
         }
 
         [Fact]
         public void GetTestNetNetworkRetuirnsNetworkTest()
         {
-            Network network = WalletHelpers.GetNetwork("testnet");
+            Network network = NetworkHelpers.GetNetwork("testnet");
             Assert.Equal(Network.TestNet, network);
         }
 
         [Fact]
         public void GetNetworkIsCaseInsensitive()
         {
-            Network testNetwork = WalletHelpers.GetNetwork("Test");
+            Network testNetwork = NetworkHelpers.GetNetwork("Test");
             Assert.Equal(Network.TestNet, testNetwork);
 
-            Network mainNetwork = WalletHelpers.GetNetwork("MainNet");
+            Network mainNetwork = NetworkHelpers.GetNetwork("MainNet");
             Assert.Equal(Network.Main, mainNetwork);
         }
 
         [Fact]
         public void WrongNetworkThrowsArgumentException()
         {
-            var exception = Record.Exception(() => WalletHelpers.GetNetwork("myNetwork"));
+            var exception = Record.Exception(() => NetworkHelpers.GetNetwork("myNetwork"));
             Assert.NotNull(exception);
             Assert.IsType<ArgumentException>(exception);
         }
