+++ src/Stratis.Bitcoin/Utilities/JsonConverters/DateTimeOffsetConverter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace Stratis.Bitcoin.Features.Wallet.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     /// <summary>
     /// Converter used to convert <see cref="DateTimeOffset"/> to and from Unix time.
