+++ src/Stratis.Bitcoin/Utilities/JsonConverters/NetworkConverter.cs
@@ -1,30 +1,31 @@
 ﻿using System;
+using NBitcoin;
 using Newtonsoft.Json;
 
-namespace Stratis.Bitcoin.Features.Wallet.JsonConverters
+namespace Stratis.Bitcoin.Utilities.JsonConverters
 {
     /// <summary>
-    /// Converter used to convert <see cref="byte"/> arrays to and from JSON.
+    /// Converter used to convert <see cref="Network"/> to and from JSON.
     /// </summary>
     /// <seealso cref="Newtonsoft.Json.JsonConverter" />
-    public class ByteArrayConverter : JsonConverter
+    public class NetworkConverter : JsonConverter
     {
         /// <inheritdoc />
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(byte[]);
+            return objectType == typeof(Network);
         }
 
         /// <inheritdoc />
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
-            return Convert.FromBase64String((string)reader.Value);
+            return NetworkHelpers.GetNetwork((string)reader.Value);
         }
 
         /// <inheritdoc />
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            writer.WriteValue(Convert.ToBase64String((byte[])value));
+            writer.WriteValue(((Network)value).ToString());
         }
     }
 }
